Add theme Blowfish v2.92.0
src: https://github.com/nunocoracao/blowfish/releases/tag/v2.92.0
We do it this way, as Blowfish has a lot of git churn and as a submodule, the version history bloats our local repo size from around 30mb to 600mb. Blowfish as a git submodule isn't much easier to update and comes with git submodule headaches. We also avoid Hugo modules, because it is overly complicated and has Golang as an additional dependency.